pyproject.toml
[project]
name = "eeg-nn"
version = "0.1.0"
description = "Create a Neural Network for MI Classification"
authors = [
    {name = "Cyril Saju"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc (>=3.66.0,<4.0.0)",
    "moabb (>=1.4.3,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "mlflow (>=3.8.1,<4.0.0)",
    "optuna (>=4.6.0,<5.0.0)",
    "mne (>=1.11.0,<2.0.0)",
    "snntorch (>=0.9.4,<0.10.0)",
    "spikingjelly (>=0.0.0.0.14,<0.0.1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}
torchvision = {source = "pytorch-gpu"}
torchaudio = {source = "pytorch-gpu"}

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)"
]
